# labs/lab03-rest-api/starter/openapi.yaml
openapi: 3.0.3
info:
  title: Lab 3 Item API
  version: 1.0.0
  description: TODO - briefly describe this API.

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /items:
    get:
      summary: List items
      description: Returns a small list of example items.
      responses:
        "200":
          description: Item list response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
              example:
                items:
                  - id: 1
                    name: keyboard
                  - id: 2
                    name: mouse
                  - id: 3
                    name: monitor

    post:
      summary: Create a new item
      description: Adds a unique item to the inventory system.
      operationId: createItem
      requestBody:
        description: The item details to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '201':
          description: Item successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid input data provided.


  /items/{id}:
    get:
      summary: Get item by ID
      description: Returns one example item using the ID from the path.
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID from the URL path.
          schema:
            type: integer
            example: 7
      responses:
        "200":
          description: Single item response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                id: 7
                name: Item 7
                quantity: 42
    put:
      summary: Update an item
      description: Replaces an existing item by its ID.
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID from the URL part
          schema:
            type: integer
            example: 7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Item_Name"
                quantity:
                  type: integer
                  example: 7
      responses:
        "200":
          description: Single item response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                id: 7
                name: Item 7
                quantity: 42
        '404':
          description: Item not found
        '400':
          description: Invalid data or malformed request

    delete:
      summary: Delete an item
      description: Deletes a specific item by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the item to delete
          schema:
            type: integer
            example: 7
      responses:
        '201':
          description: Item successfully deleted
        '404':
          description: Item not found


components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required:
        - status

    HelloResponse:
      type: object
      properties:
        message:
          type: string
          example: Hello from Express
      required:
        - message

    RequestCountResponse:
      type: object
      properties:
        count:
          type: integer
          example: 3
      required:
        - count

    EchoResponse:
      type: object
      properties:
        youSent:
          type: object
          additionalProperties: true
      required:
        - youSent

    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: keyboard
        quantity:
          type: integer
          example: 10
      required:
        - id
        - name
        - quantity

    ItemInput:
      type: object
      properties:
        name:
          type: string
          example: keyboard
        quantity:
          type: integer
          example: 10
      required:
        - name
        - quantity

    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
      required:
        - items

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Not found
      required:
        - error